+++ src/randomizer/mod.rs
@@ -153,9 +153,9 @@ fn randomize_inner(
         rom.set_tag("palettes");
         let mut palette_rng = ChaCha8Rng::from_os_rng();
         if options.palette_themed {
-            randomize::palettes::randomize_themed(rom, &mut palette_rng);
+            randomize::palettes::randomize_themed(rom, &mut palette_rng, options.player_color);
         } else {
-            randomize::palettes::randomize(rom, &mut palette_rng);
+            randomize::palettes::randomize(rom, &mut palette_rng, options.player_color);
         }
     }
     if options.any_enemies_active() {